Do not skip test utils
The OTE logic calls ModulesTestsOnly() when no filter is
passed. This filter does a simple exclusion of any
spec that is located inside /vendor/.

The openshift/origin module contains the extended util
package that is used by a sort of tests, and some of these
helper functions have a call to Ginkgo BeforeEach/AfterEach,
causing them to be recognized as part of the test.

Because the extended utils is located on a different repo, they
appear as /vendor and the whole test is ignored, as it contains some
nodes that are on /vendor directory.

This change modifies the ModuleTestsOnly to exclude any vendored
package/test, unless they are part of the openshift/origin
extended test util packages, in which case it should be part of the test